+++ src/components/QuizContainerNew/quizUtils/index.tsx
@@ -22,3 +22,16 @@ export const shortAnswerTitleFormatter = (
 export const removeMarkdown = (title: string | null | undefined): string => {
   return title ? title.replace(/\*{1,2}(.*?)\*{1,2}/g, "$1") : "";
 };
+
+export const constrainHeight = (h?: number) =>
+  h ? Math.max(Math.min(200, h), 96) : undefined;
+
+export const calcDims = (w?: number, h?: number) => {
+  const constrainedHeight = constrainHeight(h);
+  return w && h && constrainedHeight
+    ? {
+        width: Math.round((w / h) * constrainedHeight),
+        height: constrainedHeight,
+      }
+    : { width: undefined, height: undefined };
+};