+++ src/node-lib/curriculum-api-2023/queries/curriculumUnits/curriculumUnits.schema.ts
@@ -16,6 +16,7 @@ const curriculumUnitsSchema = z.object({
       lessons: z
         .array(
           z.object({
+            slug: z.string().optional(),
             title: z.string(),
             slug: z.string().optional(),
           }),